feat: add loadFriendProfile function
