+++ frontend/index.html
@@ -15,6 +15,7 @@
     <link rel="stylesheet" href="/static/css/menu/hotbar.css">
     <link rel="stylesheet" href="/static/css/menu/stats_content.css">
     <link rel="stylesheet" href="/static/css/menu/friends_content.css">
+    <link rel="stylesheet" href="/static/css/menu/friend_profile.css">
     <link rel="stylesheet" href="/static/css/menu/how_to_play_content.css">
     <link rel="stylesheet" href="/static/css/menu/play_content.css">
     <link rel="stylesheet" href="/static/css/menu/settings_content.css">
