+++ frontend/src/components/menu/chat_interface.html
@@ -14,14 +14,16 @@
   <div class='avatar-section'>
     <div class='avatar-container'>
       <img src='/static/images/kirby.png' alt='avatar' class='profile-settings-avatar avatar'>
-      <!-- <div class='status-badge'></div> -->
+      <div class='status-badge'></div>
     </div>
-    rsoo
+    <div class='username'>
+      username
+    </div>
   </div>
   <div class='pvp-stat-container'>
-    <span>rsoo vs melee</span>
+    <span class='pvp-users-label'>you vs friend</span>
     <div class='win-lose-stat-bar'>
-      <span>5 - 5</span>
+      <span class='pvp-tally-label'>0 - 0</span>
       <div class='lose-bar'>
         <div class='win-bar'></div>
       </div>
@@ -34,7 +36,7 @@
   <div class='total-game-stat-container'>
     <span>Total Games</span>
     <div class='win-lose-stat-bar'>
-      <span>5 - 5</span>
+      <span class='total-tally-label'>0 - 0</span>
       <div class='lose-bar'>
         <div class='win-bar'></div>
       </div>
