+++ frontend/src/css/menu/friend_profile.css
@@ -1,0 +1,78 @@
+#profile-container {
+  padding-top: 2rem;
+  width: 13rem;
+  height: 100%;
+  border-radius: 1rem;
+  background-color: var(--charcoal-800);
+  box-sizing: border-box;
+  display: flex;
+  flex-direction: column;
+  gap: 2rem;
+}
+
+#profile-container > .avatar-section {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  font-size: 1.5rem;
+}
+
+#profile-container .avatar-container {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+  height: 100%;
+  position: relative;
+}
+
+#profile-container .avatar {
+  width: 6rem;
+  height: 6rem;
+}
+
+#profile-container .status-badge {
+  width: 1.25rem;
+  height: 1.25rem;
+  position: absolute;
+  bottom: 0.05rem;
+  right: 3.75rem;
+  border-width: 0.2rem;
+}
+
+.pvp-stat-container,
+.total-game-stat-container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 0.3125rem;
+}
+
+.win-lose-stat-bar {
+  display: flex;
+  flex-direction: column;
+  gap: 0.2rem;
+  width: 10rem;
+  align-items: center;
+}
+
+.lose-bar {
+  width: 100%;
+  height: 0.5rem;
+  background-color: var(--magenta-500);
+  border-radius: 0.5rem;
+}
+
+.win-bar {
+  width: 50%;
+  height: 0.5rem;
+  background-color: var(--teal-500);
+  border-radius: 0.5rem 0 0 0.5rem;
+}
+
+.win-lose-label {
+  display: flex;
+  justify-content: space-between;
+  min-width: 10rem;
+}
+
