+++ frontend/src/css/menu/friends_content.css
@@ -325,80 +325,6 @@
   font-weight: 600;
 }
 
-/*profile-container*/
-#profile-container {
-  padding-top: 2rem;
-  width: 13rem;
-  height: 100%;
-  border-radius: 1rem;
-  background-color: var(--charcoal-800);
-  box-sizing: border-box;
-  display: flex;
-  flex-direction: column;
-  gap: 2rem;
-}
-
-#profile-container > .avatar-section {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  font-size: 1.5rem;
-}
-
-#profile-container .avatar-container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-  height: 100%;
-}
-
-#profile-container .avatar {
-  width: 6rem;
-  height: 6rem;
-}
-
-#profile-container .status-badge {
-  width: 1.6rem;
-  height: 1.6rem;
-}
-
-.pvp-stat-container,
-.total-game-stat-container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 0.3125rem;
-}
-
-.win-lose-stat-bar {
-  display: flex;
-  flex-direction: column;
-  gap: 0.2rem;
-  width: 10rem;
-  align-items: center;
-}
-
-.lose-bar {
-  width: 100%;
-  height: 0.5rem;
-  background-color: var(--magenta-500);
-  border-radius: 0.5rem;
-}
-
-.win-bar {
-  width: 50%;
-  height: 0.5rem;
-  background-color: var(--teal-500);
-  border-radius: 0.5rem 0 0 0.5rem;
-}
-
-.win-lose-label {
-  display: flex;
-  justify-content: space-between;
-  min-width: 10rem;
-}
-
 #send-message-button {
   background-color: transparent;
   border: none;
